requirements.txt
django
gunicorn
django-heroku
djangorestframework
Pillow